fix stdout buffer
