+++ src/html2svg.ts
@@ -1,85 +1,65 @@
-import { program } from 'commander'
 import { app, BrowserWindow } from 'electron'
 
-const entry = process.argv.find((a) => a.endsWith('html2svg.js'))
-const index = entry ? process.argv.indexOf(entry) : -1
-const args = process.argv.slice(Math.max(2, index + 1))
+export interface Options {
+    full?: boolean
+    wait?: number
+    width?: number
+    height?: number
+    format?: 'svg' | 'pdf' | 'png' | 'jpg' | 'webp'
+}
 
-program
-    .name('html2svg')
-    .showHelpAfterError()
-    .showSuggestionAfterError()
-    .argument('<url>', 'URL to the web page to render')
-    .option('-f, --full', 'capture the entire page')
-    .option(
-        '-w, --wait <seconds>',
-        'set the amount of seconds to wait between the page loaded event and taking the screenshot',
-        validateInt,
-        1,
-    )
-    .option(
-        '-w, --width <width>',
-        'set the viewport width in pixels',
-        validateInt,
-        1920,
-    )
-    .option(
-        '-h, --height <height>',
-        'set the viewport height in pixels',
-        validateInt,
-        1080,
-    )
-    .option(
-        '-f, --format <format>',
-        'set the output format, should one of these values: svg, pdf, png, jpg, webp',
-        'svg',
-    )
-    .action(async (url, { full, wait, width, height, format }) => {
-        const mode = getMode(format)
+app.dock?.hide()
+app.disableHardwareAcceleration()
+app.commandLine.appendSwitch('headless')
+app.commandLine.appendSwitch('no-sandbox')
+app.commandLine.appendSwitch('disable-gpu')
 
-        if (format === 'svg') {
-            process.env.html2svg_svg_mode = 'true'
-        }
+export async function html2svg(
+    url: string,
+    { full, wait, width, height, format }: Options = {},
+) {
+    const mode = getMode(format ?? 'svg')
 
-        app.dock?.hide()
-        app.disableHardwareAcceleration()
-        app.commandLine.appendSwitch('headless')
-        app.commandLine.appendSwitch('no-sandbox')
-        app.commandLine.appendSwitch('disable-gpu')
+    await app.whenReady()
 
-        await app.whenReady()
+    const page = new BrowserWindow({
+        width,
+        height,
+        show: false,
+        webPreferences: {
+            sandbox: false,
+            offscreen: true,
+            additionalArguments:
+                format === 'svg'
+                    ? ['--html2svg-svg-mode', '--disable-remote-fonts']
+                    : [],
+        },
+    })
 
-        const page = new BrowserWindow({
-            width,
-            height,
-            show: false,
-            webPreferences: { sandbox: false },
-        })
+    try {
+        await new Promise<void>((resolve, reject) =>
+            Promise.resolve()
+                .then(async () => {
+                    const timeout = setTimeout(() => {
+                        page.webContents.off('did-finish-load', listener)
 
-        try {
-            await new Promise<void>((resolve, reject) =>
-                Promise.resolve()
-                    .then(async () => {
-                        const timeout = setTimeout(() => {
-                            page.webContents.off('did-finish-load', listener)
+                        reject(new Error('timeout'))
+                    }, 10_000)
+                    const listener = () => {
+                        clearTimeout(timeout)
 
-                            reject(new Error('timeout'))
-                        }, 10_000)
-                        const listener = () => {
-                            clearTimeout(timeout)
+                        resolve()
+                    }
 
-                            resolve()
-                        }
+                    page.webContents.once('did-finish-load', listener)
 
-                        page.webContents.once('did-finish-load', listener)
+                    await page.loadURL(url)
+                })
+                .catch(reject),
+        )
 
-                        await page.loadURL(url)
-                    })
-                    .catch(reject),
-            )
-
-            await page.webContents.executeJavaScript(
-                `
+        await page.webContents.executeJavaScript(
+            `
                     new Promise(resolve => {
                         const style = document.createElement('style')
 
@@ -100,7 +80,7 @@ program
                             scrollTo({ top: 0 })
 
                             requestAnimationFrame(() =>
-                                setTimeout(resolve, ${wait * 1000})
+                                setTimeout(resolve, ${(wait ?? 0) * 1000})
                             )
                         })
                     }).then(() =>
@@ -111,19 +91,11 @@ program
                         )
                     )
                 `,
-            )
-        } finally {
-            page.destroy()
-        }
-
-        process.exit(0)
-    })
-    .parseAsync(args, { from: 'user' })
-    .catch((error) => {
-        console.error(error)
-
-        process.exit(1)
-    })
+        )
+    } finally {
+        page.destroy()
+    }
+}
 
 function getMode(format: string) {
     switch (format) {
@@ -141,14 +113,4 @@ function getMode(format: string) {
         default:
             throw new Error(`Unsupported output format: ${format}`)
     }
-}
-
-function validateInt(string: string) {
-    const number = parseInt(string, 10)
-
-    if (Number.isNaN(number)) {
-        throw new Error(`Invalid number value: ${string}`)
-    }
-
-    return number
 }
